\name{extractoR.description-package}
\alias{extractoR.description-package}
\alias{extractoR.description}
\docType{package}
\encoding{UTF-8}
\title{R packages metadata extractor (extract)}
\description{
  extractoR.description is part of extractoR which is a set of R
  packages that can be used to fetch, extract and query R packages
  metadata. extractoR.description parses DESCRIPTION files from R
  packages.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}
