\name{MakeGlobalIndex}
\alias{MakeGlobalIndex}
\encoding{UTF-8}
\title{MakeGlobalIndex}
\description{
  Global Index.
}

\usage{
MakeGlobalIndex(datadir, sources=c("cran", "github"))
}
\arguments{
  \item{datadir}{Directory where to store and extract the CRAN and
    Github indexes.}
  \item{sources}{Package sources to use.}
}
\value{
  The global CRAN and Github R package index.
}

\author{
  Maëlick Claes <himself@maelick.net>
}

\keyword{package}
\keyword{index}
\keyword{cran}
\keyword{github}
